requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.6
djangorestframework==3.15.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
